Setup the project to be easier to reason about
