--- man/man.go
@@ -1,17 +1,0 @@
-package man
-
-// These are messages used when running help commands such as --help or -h
-
-const ROOT_HELP_MENU = `Looks is a tool for artists to have easier access to generative art.
-
-Usage:
-
-	looks <command> [arguments] [flags]
-
-Commands:
-
-	generate	Generates new pieces from given assets
-
-Flags:
-
-	--no-meta  skip generating metadata`
